Add battery check service and update deployment configurations

+++ lib/domain/usecases/reboot.dart
@@ -1,7 +1,11 @@
 import 'package:stpvelox/core/utils/sudo_process.dart';
 
 class RebootDevice {
-  Future<void> call() async {
-    await SudoProcess.run('reboot', []);
+  Future<void> call([bool isShutdown = false]) async {
+    if (isShutdown) {
+      await SudoProcess.run('shutdown', [' -h', 'now']);
+    } else {
+      await SudoProcess.run('reboot', []);
+    }
   }
 }
